activebike_khv/requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.6.0
colorama==0.4.4
Django==3.2.9
django-debug-toolbar==3.7.0
django-menu==0.1.13
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-isort==4.1.1
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
gpxpy==1.5.0
hashring==1.5.1
importlib-metadata==5.0.0
iniconfig==1.1.1
isort==5.10.1
Markdown==3.4.1
mccabe==0.6.1
memcache==0.12.0
mysqlclient==2.1.0
packaging==21.3
pep8-naming==0.12.1
Pillow==9.1.0
pluggy==1.0.0
polyline==1.4.0
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.6
pytest==6.2.5
python-decouple==3.6
pytz==2021.3
rsa==4.9
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.2
Telethon==1.25.4
testfixtures==6.18.3
toml==0.10.2
zipp==3.10.0
